requirements.txt
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
dj-database-url==1.2.0
Django==4.1.4
djangorestframework==3.14.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
ipykernel==6.19.2
ipython==8.7.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.8
jupyter_core==5.1.0
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.0
openai==0.25.0
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.0
plotly==5.11.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pyzmq==24.0.1
requests==2.28.1
scikit-learn==1.2.0
scipy==1.9.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
tenacity==8.1.0
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
